Use accessor methods instead of accessing structure directly.
This sounds good on paper but seems kind of overly anal to me whe
programming in C. Especially when I am given the only copy
and I am expected to free the memory associated with the data
structure. Oh well!

Removes 'const' qualifiers from convert_ function parameters. We
are not given const pointers to structures and I would have to cast
stuff or something.