Rewrite ALPM::LoadConfig. Rename to ALPM::Conf.
Test has been rewritten. Typemap is adjusted to allow for string
options that have not yet been set. Adjust the MANIFEST for the
renamed file and the new lib/ALPM/Conf.pod file where the POD
has been moved from within the .pm file.

I need to rethink how to make pacman.conf file parsing as easy as
it was on the command line. Maybe it will be started from the
ALPM::Conf module instead. For example:
perl -MALPM::Conf=/etc/pacman.conf -e 'print $alpm->get_rootdir'

In order to set the rootdir from the conf file I also might
need to rewrite the parser so that it stores all of the option values
and then spawns an alpm handle.