Create import magic in ALPM::Conf for quick scripting.
Previously I had a special import sub in ALPM that allowed one to quickly
load a pacman.conf file. Now that ALPM::Conf creates a new ALPM handle,
this is more easily accomplished from the ALPM::Conf module. When
"use ALPM::Conf" is given an argument it will treat the argument as the
path to a conf file. The conf file is parsed and a new alpm handle is
imported into the caller's namespace.

This is handy for embedding perl in shell scripts or the command line, i.e.
perl -MALPM::Conf=/etc/pacman.conf -E '
for $p ($alpm->localdb->pkgs){
	say $p->name;
}
'