bless objects into packages
This provides interfaces that insulate users from changes in how the
underlying object is stored, allowing for a smoother transition to using
the raw C struct pointer in the future.  ALPM::Dependency and
ALPM::FileList in particular are likely candidates for such a conversion
because alpm provides functions that take these types as input.  This
will also allow stringification methods to be provided for easier use in
messages.